+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Revision history for MetaLambda
 
 ## Dev
+Add
+* `Subst` type (simultaneous substitution)
+
 Change
 * let-box syntax `let box[Ψ.U] = e1 in e2` -> `let box[U] = e1 in e2`
 * clo syntax `U with (e1, ..., en)` -> `U with [e1/x1, ..., en/xn]`